+++ Data/Datatest.tsx
@@ -1,0 +1,38 @@
+export const sampleRequests = [
+    {
+      id: 1,
+      type: 'maintenance',
+      title: 'Reparación de aire acondicionado - Sala 202',
+      description: 'El aire acondicionado no enfría adecuadamente',
+      priority: 'high',
+      status: 'pending',
+      requestedBy: 'Dr. María González',
+      department: 'Cardiología',
+      createdAt: '2024-08-10T09:30:00',
+      updatedAt: '2024-08-10T09:30:00'
+    },
+    {
+      id: 2,
+      type: 'tech_support',
+      title: 'Computadora no enciende - Recepción',
+      description: 'La computadora principal de recepción no responde',
+      priority: 'urgent',
+      status: 'in_progress',
+      requestedBy: 'Ana López',
+      department: 'Administración',
+      createdAt: '2024-08-09T14:15:00',
+      updatedAt: '2024-08-10T08:00:00'
+    },
+    {
+      id: 3,
+      type: 'maintenance',
+      title: 'Filtración en baño - Piso 3',
+      description: 'Hay una pequeña filtración en el baño del tercer piso',
+      priority: 'medium',
+      status: 'completed',
+      requestedBy: 'Carlos Ramírez',
+      department: 'Mantenimiento',
+      createdAt: '2024-08-08T11:00:00',
+      updatedAt: '2024-08-09T16:30:00'
+    }
+  ];
