+++ app/layout.tsx
@@ -1,10 +1,13 @@
 import { FC, PropsWithChildren } from "react";
 import "./ui/globals.css";
+import { Header } from "@/components/layout/header";
 
-const layout:FC<PropsWithChildren> = ({children}) => {
+
+const layout: FC<PropsWithChildren> = ({ children }) => {
   return (
     <html>
       <body>
+        <Header />
         {children}
       </body>
     </html>
