+++ app/login/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+const Login = () => {
+    return (
+        <div>
+            <h1>Login</h1>
+        </div>
+    )
+}
+export default Login;
